Fetch, parse, & print weather report(s) from .gov
Use two XML API endpoints at weather.gov. One accepts a station code to
retrieve the most recent weather report from that station. The other
retrieves the most recent report from all stations.

If all reports were fetched, the payload is zipped. Unzip the payload
in memory, and select the reports that are mostly likely to be valid,
current station reports.

If all reports were fetched and the --page and --per-page flags were
passed in, then return one "page" of results from the full result set.
This pagination is done on our side, in memory, from the full result
set, because the endpoint doesn't support pagination.

Extract some of the data from the downloaded report(s) by using either
of two preconfigured lists of fields as keys to extract their values
from the XML. The default is the shorter of the two preconfigured
lists; --full can be passed to use the longer list and get a wider
table.

Align the data in tabular format by getting, for each field, the length
of the longest string from all rows (including the headers) for that
field. Pad the feader and each value for that field to that length.